pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "personalized_chatbot"
version = "0.1.0"
description = "Personalized, emotion-aware conversational assistant for mental health support"
requires-python = ">=3.10"
dependencies = [
    "flask>=3.0.0",
    "pydantic>=2.7.0",
    "python-dotenv>=1.0.0",
    "transformers>=4.44.0",
    "torch>=2.1.0",
    "sentence-transformers>=3.0.0",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "onnxruntime>=1.18.0",
    "google-generativeai>=0.7.2",
    "opencv-python>=4.9.0",
    "fer>=22.5.0",
    "sqlalchemy>=2.0.30",
    "openai-whisper>=20240930",
    "chromadb>=0.4.0"
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
addopts = "-ra"
testpaths = ["tests"]